Refactor createSystemCalls
